Version 0.0.2 in config.xml, more changes
Set properties name case sensitive
Also, as we need build a version with portrait orientation
for the tablets, and one with landscape orientation for the xo-4
(because the former use on-screen keyboard, and the second use
a physical keyboard), add a comment.
This change means we need build once for tablets, change the config.xml
and build for the xo-4. I didn't find other way.